requirements.txt
openai
streamlit
python-dotenv
sqlalchemy
pandas
datetime
python-docx
webvtt-py